LoopManager::mStoredCloud become a float value

+++ sources/gameplay/loopmanager.hpp
@@ -43,7 +43,7 @@ private:
     std::vector< GameEntity* > mEntities;
     std::vector<std::shared_ptr<Texture2D>> mCloudsTextures;
     int mCloudTextureIdx = 0;
-    int mStoredCloud = 0;
+    float mStoredCloud = 0.f;
     bool mClickLeft = false;
     bool mShiftLeft = false;
     bool mCtrlLeft = false;
